Modularized source code
- modularized source code
- installed react dev tools
