+++ .rescriptsrc.js
@@ -1,4 +1,4 @@
-module.exports = config => {
+module.exports = (config) => {
     config.target = 'electron-renderer';
     return config;
 }
