+++ src/main.js
@@ -3,6 +3,8 @@ const isDev = require('electron-is-dev');
 const path = require('path');
 let mainWindow;
 
+const { default: installExtension, REACT_DEVELOPER_TOOLS } = require('electron-devtools-installer');
+
 const createWindow = () => {
     mainWindow = new BrowserWindow({
         width: 800,
@@ -20,7 +22,14 @@ const createWindow = () => {
     mainWindow.once('ready-to-show', () => mainWindow.show());
 }
 
-app.on('ready', createWindow);
+app.on('ready', () => {
+    installExtension(REACT_DEVELOPER_TOOLS).then((name) => {
+        console.log(`Loaded Extension:  ${name}`);
+    }).catch((err) => {
+        console.log('An error occurred installing the React Developer Tools: ', err);
+    });
+    createWindow();
+});
 app.on('window-all-closed', () => {
     if (process.platform !== 'darwin') {
         app.quit();
